requirements.txt
notebook
scipy
numpy
matplotlib
seaborn
pandas
geopandas
scikit-learn
fiona
Shapely
pyproj
plotly
rasterio
scikit-gstat
statsmodels
flopy
pyshp==2.3.0
plpygis
python-dotenv
pysheds
pykrige
rasterstats
osmnx
pyemu
jupyter_contrib_nbextensions
contextily
PyYAML
pydantic
